fix: fix import path issue after renaming
