+++ src/commands/deploy.ts
@@ -1,6 +1,6 @@
 import { deployStack } from '../stack';
 import { constructActionContext, logger } from '../common';
-import { parseYaml } from '../parses';
+import { parseYaml } from '../parser';
 
 export const deploy = async (
   stackName: string,
